+++ server/src/main/java/org/opensearch/tasks/TaskCancellation.java
@@ -9,6 +9,7 @@
 package org.opensearch.tasks;
 
 import org.opensearch.ExceptionsHelper;
+import org.opensearch.rest.RestStatus;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,11 +27,13 @@ public class TaskCancellation implements Comparable<TaskCancellation> {
     private final CancellableTask task;
     private final List<Reason> reasons;
     private final List<Runnable> onCancelCallbacks;
+    private final RestStatus restStatus;
 
-    public TaskCancellation(CancellableTask task, List<Reason> reasons, List<Runnable> onCancelCallbacks) {
+    public TaskCancellation(CancellableTask task, List<Reason> reasons, List<Runnable> onCancelCallbacks, RestStatus restStatus) {
         this.task = task;
         this.reasons = reasons;
         this.onCancelCallbacks = onCancelCallbacks;
+        this.restStatus = restStatus;
     }
 
     public CancellableTask getTask() {
@@ -53,7 +56,7 @@ public class TaskCancellation implements Comparable<TaskCancellation> {
             return;
         }
 
-        task.cancel(getReasonString());
+        task.cancel(new CancellableTask.Reason(getReasonString(), restStatus));
 
         List<Exception> exceptions = new ArrayList<>();
         for (Runnable callback : onCancelCallbacks) {